JIRA: https://stackstorm.atlassian.net/browse/CANDY-127
As an administrator, need the ability to delete an L2 or L3 ACL on a switch so that a I can allow or deny certain type of traffic.
Must support following input parameters, required parameters are in bold.
mgmt_ip: ip_type
acl_name: string
type : switched | routed
Output parameters:
TRUE for Success, FALSE for Failure